+++ src/rendering/measurefragment.ts
@@ -1,19 +1,31 @@
-import { Config } from './config';
-import { Stave, StaveModifier, StaveRendering } from './stave';
 import * as musicxml from '@/musicxml';
 import * as util from '@/util';
+import * as vexflow from 'vexflow';
+import { Config } from './config';
 import { MeasureEntry, StaveSignature } from './stavesignature';
+import { PartScoped } from './types';
 import { Address } from './address';
+import { Part, PartRendering } from './part2';
+import { Chorus, ChorusRendering } from './chorus';
 import { Spanners } from './spanners';
+import { StaveModifier } from './stave';
 
 const STAVE_SIGNATURE_ONLY_MEASURE_FRAGMENT_PADDING = 8;
+const STAVE_CONNECTOR_BRACE_WIDTH = 16;
 
 /** The result of rendering a measure fragment. */
 export type MeasureFragmentRendering = {
   type: 'measurefragment';
   address: Address<'measurefragment'>;
-  staves: StaveRendering[];
+  parts: PartRendering[];
   width: number;
+};
+
+/** The width of a measure fragment. */
+export type MeasureFragmentWidth = {
+  measureIndex: number;
+  measureFragmentIndex: number;
+  value: number;
 };
 
 /**
@@ -21,38 +33,42 @@ export type MeasureFragmentRendering = {
  *
  * A measure fragment is necessary when stave modifiers change. It is not a formal musical concept, and it is moreso an
  * outcome of vexflow's Stave implementation.
+ *
+ * Measure fragments format all measure parts against the first stave.
  */
 export class MeasureFragment {
   private config: Config;
   private index: number;
-  private leadingStaveSignature: StaveSignature;
-  private measureEntries: MeasureEntry[];
-  private staveLayouts: musicxml.StaveLayout[];
-  private staveCount: number;
-  private beginningBarStyle: musicxml.BarStyle;
-  private endBarStyle: musicxml.BarStyle;
+  private partIds: string[];
+  private musicXml: {
+    staveLayouts: musicxml.StaveLayout[];
+    beginningBarStyles: PartScoped<musicxml.BarStyle>[];
+    endBarStyles: PartScoped<musicxml.BarStyle>[];
+  };
+  private measureEntries: PartScoped<MeasureEntry>[];
+  private staveSignatures: PartScoped<StaveSignature>[];
 
   constructor(opts: {
     config: Config;
     index: number;
-    leadingStaveSignature: StaveSignature;
-    measureEntries: MeasureEntry[];
-    staveLayouts: musicxml.StaveLayout[];
-    staveCount: number;
-    beginningBarStyle: musicxml.BarStyle;
-    endBarStyle: musicxml.BarStyle;
+    partIds: string[];
+    musicXml: {
+      staveLayouts: musicxml.StaveLayout[];
+      beginningBarStyles: PartScoped<musicxml.BarStyle>[];
+      endBarStyles: PartScoped<musicxml.BarStyle>[];
+    };
+    measureEntries: PartScoped<MeasureEntry>[];
+    staveSignatures: PartScoped<StaveSignature>[];
   }) {
     this.config = opts.config;
     this.index = opts.index;
-    this.leadingStaveSignature = opts.leadingStaveSignature;
+    this.partIds = opts.partIds;
+    this.musicXml = opts.musicXml;
     this.measureEntries = opts.measureEntries;
-    this.staveLayouts = opts.staveLayouts;
-    this.staveCount = opts.staveCount;
-    this.beginningBarStyle = opts.beginningBarStyle;
-    this.endBarStyle = opts.endBarStyle;
+    this.staveSignatures = opts.staveSignatures;
   }
 
-  /** Returns the index of the measure fragment within the measure. */
+  /** Returns the index of the measure fragment, which is relative to its parent measure. */
   getIndex(): number {
     return this.index;
   }
@@ -62,52 +78,34 @@ export class MeasureFragment {
     address: Address<'measurefragment'>;
     previousMeasureFragment: MeasureFragment | null;
   }): number {
-    const staveModifiers = this.getStaveModifiers({
-      address: opts.address,
-      previousMeasureFragment: opts.previousMeasureFragment,
-    });
-    const staveModifiersWidth = this.getStaveModifiersWidth(Array.from(staveModifiers));
+    const address = opts.address;
 
-    return this.getMinJustifyWidth(opts.address) + staveModifiersWidth + this.getRightPadding();
+    return (
+      this.getStaveModifiersWidth({ address, previousMeasureFragment: opts.previousMeasureFragment }) +
+      this.getMinVoiceJustifyWidth({ address }) +
+      this.getLeftPadding({ address }) +
+      this.getRightPadding()
+    );
   }
 
-  /** Returns the top padding for the measure fragment. */
+  /** Returns the top padding of the fragment. */
   getTopPadding(): number {
-    return util.max(this.getStaves().map((stave) => stave.getTopPadding()));
+    return util.max(this.getParts().map((part) => part.getTopPadding()));
   }
 
-  getMultiRestCount(): number {
-    // TODO: One stave could be a multi measure rest, while another one could have voices.
-    return util.max(this.getStaves().map((stave) => stave.getMultiRestCount()));
-  }
-
-  /** Renders the MeasureFragment. */
+  /** Renders the measure fragment. */
   render(opts: {
     x: number;
     y: number;
     address: Address<'measurefragment'>;
     spanners: Spanners;
-    systemCount: number;
-    targetSystemWidth: number;
-    minRequiredSystemWidth: number;
+    width: MeasureFragmentWidth;
     previousMeasureFragment: MeasureFragment | null;
     nextMeasureFragment: MeasureFragment | null;
   }): MeasureFragmentRendering {
-    const staveRenderings = new Array<StaveRendering>();
+    const partRenderings = new Array<PartRendering>();
 
-    const isLastSystem = opts.address.getSystemIndex() === opts.systemCount - 1;
-    const width = isLastSystem
-      ? this.getMinRequiredWidth({
-          address: opts.address,
-          previousMeasureFragment: opts.previousMeasureFragment,
-        })
-      : this.getSystemFitWidth({
-          address: opts.address,
-          minRequiredSystemWidth: opts.minRequiredSystemWidth,
-          targetSystemWidth: opts.targetSystemWidth,
-          previousMeasureFragment: opts.previousMeasureFragment,
-        });
-
+    const x = opts.x;
     let y = opts.y;
 
     const staveModifiers = this.getStaveModifiers({
@@ -115,105 +113,103 @@ export class MeasureFragment {
       previousMeasureFragment: opts.previousMeasureFragment,
     });
 
-    // Render staves.
-    util.forEachTriple(this.getStaves(), ([previousStave, currentStave, nextStave], { isFirst, isLast }) => {
-      if (isFirst) {
-        previousStave = util.last(opts.previousMeasureFragment?.getStaves() ?? []);
-      }
-      if (isLast) {
-        nextStave = util.first(opts.nextMeasureFragment?.getStaves() ?? []);
-      }
+    const vfFormatter = new vexflow.Formatter();
 
-      const staveRendering = currentStave.render({
-        x: opts.x,
+    const previousParts = opts.previousMeasureFragment?.getParts() ?? [];
+    const currentParts = this.getParts();
+    const nextParts = opts.nextMeasureFragment?.getParts() ?? [];
+
+    for (let partIndex = 0; partIndex < currentParts.length; partIndex++) {
+      const previousPart = previousParts[partIndex] ?? null;
+      const currentPart = currentParts[partIndex];
+      const nextPart = nextParts[partIndex] ?? null;
+
+      const partRendering = currentPart.render({
+        x,
         y,
-        address: opts.address.stave({ staveNumber: currentStave.getNumber() }),
+        vexflow: { formatter: vfFormatter },
+        address: opts.address.part({ partId: currentPart.getId() }),
         spanners: opts.spanners,
-        width,
-        modifiers: staveModifiers,
-        previousStave,
-        nextStave,
+        nextPart,
+        previousPart,
+        staveModifiers,
+        width: opts.width.value,
       });
-      staveRenderings.push(staveRendering);
 
-      const staveDistance =
-        this.staveLayouts.find((staveLayout) => staveLayout.staveNumber === staveRendering.staveNumber)
-          ?.staveDistance ?? this.config.DEFAULT_STAVE_DISTANCE;
+      partRenderings.push(partRendering);
 
-      y += staveDistance;
-    });
+      y += partRendering.height + this.config.PART_DISTANCE;
+    }
+
+    const vfStave = util.first(partRenderings)?.staves[0]?.vexflow.stave ?? null;
+    const vfVoices = partRenderings
+      .flatMap((partRendering) => partRendering.staves)
+      .map((stave) => stave.entry)
+      .filter((entry): entry is ChorusRendering => entry.type === 'chorus')
+      .flatMap((chorusRendering) => chorusRendering.voices)
+      .map((voice) => voice.vexflow.voice);
+    if (vfStave && vfVoices.some((vfVoice) => vfVoice.getTickables().length > 0)) {
+      vfFormatter.formatToStave(vfVoices, vfStave);
+    }
 
     return {
       type: 'measurefragment',
       address: opts.address,
-      staves: staveRenderings,
-      width,
+      parts: partRenderings,
+      width: opts.width.value,
     };
   }
 
   @util.memoize()
-  private getStaves(): Stave[] {
-    const staves = new Array<Stave>(this.staveCount);
+  private getParts(): Part[] {
+    return this.partIds.map((partId) => {
+      const measureEntries = this.measureEntries
+        .filter((measureEntry) => measureEntry.partId === partId)
+        .map((measureEntry) => measureEntry.value);
 
-    for (let staveIndex = 0; staveIndex < this.staveCount; staveIndex++) {
-      const staveNumber = staveIndex + 1;
+      const staveSignature = this.staveSignatures.find((staveSignature) => staveSignature.partId === partId)?.value;
+      if (!staveSignature) {
+        throw new Error(`Could not find stave signature for part ${partId}`);
+      }
 
-      staves[staveIndex] = new Stave({
+      const beginningBarStyle = this.musicXml.beginningBarStyles.find((barStyle) => barStyle.partId === partId)?.value;
+      if (!beginningBarStyle) {
+        throw new Error(`Could not find beginning bar style for part ${partId}`);
+      }
+
+      const endBarStyle = this.musicXml.endBarStyles.find((barStyle) => barStyle.partId === partId)?.value;
+      if (!endBarStyle) {
+        throw new Error(`Could not find end bar style for part ${partId}`);
+      }
+
+      return new Part({
         config: this.config,
-        staveSignature: this.leadingStaveSignature,
-        number: staveNumber,
+        id: partId,
         musicXml: {
-          beginningBarStyle: this.beginningBarStyle,
-          endBarStyle: this.endBarStyle,
+          staveLayouts: this.musicXml.staveLayouts,
+          beginningBarStyle,
+          endBarStyle,
         },
-        measureEntries: this.measureEntries.filter((entry) => {
-          if (entry instanceof musicxml.Note) {
-            return entry.getStaveNumber() === staveNumber;
-          }
-          return true;
-        }),
+        measureEntries,
+        staveSignature,
       });
-    }
-
-    return staves;
+    });
   }
 
-  /** Returns the minimum justify width. */
-  @util.memoize()
-  private getMinJustifyWidth(address: Address<'measurefragment'>): number {
-    return util.max(
-      this.getStaves().map((stave) => stave.getMinJustifyWidth(address.stave({ staveNumber: stave.getNumber() })))
-    );
-  }
-
-  /** Returns the right padding of the measure fragment. */
-  private getRightPadding(): number {
-    let padding = 0;
-
-    if (this.measureEntries.length === 1 && this.measureEntries[0] instanceof StaveSignature) {
-      padding += STAVE_SIGNATURE_ONLY_MEASURE_FRAGMENT_PADDING;
-    }
-
-    return padding;
-  }
-
-  /** Returns the width needed to stretch to fit the target width of the System. */
-  private getSystemFitWidth(opts: {
+  private getStaveModifiersWidth(opts: {
     address: Address<'measurefragment'>;
-    targetSystemWidth: number;
-    minRequiredSystemWidth: number;
     previousMeasureFragment: MeasureFragment | null;
   }): number {
-    const minRequiredWidth = this.getMinRequiredWidth({
+    const staveModifiers = this.getStaveModifiers({
       address: opts.address,
       previousMeasureFragment: opts.previousMeasureFragment,
     });
 
-    const widthDeficit = opts.targetSystemWidth - opts.minRequiredSystemWidth;
-    const widthFraction = minRequiredWidth / opts.minRequiredSystemWidth;
-    const widthDelta = widthDeficit * widthFraction;
-
-    return minRequiredWidth + widthDelta;
+    return util.max(
+      this.getParts()
+        .flatMap((part) => part.getStaves())
+        .map((stave) => stave.getModifiersWidth(staveModifiers))
+    );
   }
 
   /** Returns what modifiers to render. */
@@ -227,20 +223,92 @@ export class MeasureFragment {
 
     const staveModifiersChanges = new Set<StaveModifier>();
 
-    for (let staveIndex = 0; staveIndex < this.staveCount; staveIndex++) {
-      const currentStave = this.getStaves()[staveIndex];
-      const previousStave = opts.previousMeasureFragment?.getStaves()[staveIndex] ?? null;
+    for (const partId of this.partIds) {
+      const staveSignature = this.staveSignatures.find((staveSignature) => staveSignature.partId === partId)?.value;
+      if (!staveSignature) {
+        continue;
+      }
 
-      for (const staveModifier of currentStave.getModifierChanges({ previousStave })) {
-        staveModifiersChanges.add(staveModifier);
+      const staveCount = staveSignature.getStaveCount();
+
+      for (let staveIndex = 0; staveIndex < staveCount; staveIndex++) {
+        const currentStave =
+          this.getParts()
+            ?.find((part) => part.getId() === partId)
+            ?.getStaves()[staveIndex] ?? null;
+
+        const previousStave =
+          opts.previousMeasureFragment
+            ?.getParts()
+            .find((part) => part.getId() === partId)
+            ?.getStaves()[staveIndex] ?? null;
+
+        const staveModifiers = currentStave?.getModifierChanges({ previousStave }) ?? [];
+
+        for (const staveModifier of staveModifiers) {
+          staveModifiersChanges.add(staveModifier);
+        }
       }
     }
 
     return Array.from(staveModifiersChanges);
   }
 
-  /** Returns the modifiers width. */
-  private getStaveModifiersWidth(staveModifiers: StaveModifier[]): number {
-    return util.max(this.getStaves().map((stave) => stave.getModifiersWidth(staveModifiers)));
+  private getMinVoiceJustifyWidth(opts: { address: Address<'measurefragment'> }): number {
+    const spanners = new Spanners();
+    const vfFormatter = new vexflow.Formatter();
+    const vfVoices = new Array<vexflow.Voice>();
+
+    for (const part of this.getParts()) {
+      const partAddress = opts.address.part({ partId: part.getId() });
+
+      for (const stave of part.getStaves()) {
+        const entry = stave.getEntry();
+
+        let vfPartStaveVoices = new Array<vexflow.Voice>();
+
+        if (entry instanceof Chorus) {
+          const address = partAddress.stave({ staveNumber: stave.getNumber() }).chorus();
+          const chorusRendering = entry.render({ address, spanners });
+          vfPartStaveVoices = chorusRendering.voices.map((voice) => voice.vexflow.voice);
+        }
+
+        if (vfPartStaveVoices.length > 0) {
+          vfFormatter.joinVoices(vfPartStaveVoices);
+        }
+
+        vfVoices.push(...vfPartStaveVoices);
+      }
+    }
+
+    if (vfVoices.length === 0) {
+      return 0;
+    }
+
+    return vfFormatter.preCalculateMinTotalWidth(vfVoices) + spanners.getPadding() + this.config.VOICE_PADDING;
+  }
+
+  private getLeftPadding(opts: { address: Address<'measurefragment'> }): number {
+    let padding = 0;
+
+    const hasStaveConnectorBrace =
+      opts.address.getSystemMeasureIndex() === 0 &&
+      this.index === 0 &&
+      this.staveSignatures.some((staveSignature) => staveSignature.value.getStaveCount() > 1);
+    if (hasStaveConnectorBrace) {
+      padding += STAVE_CONNECTOR_BRACE_WIDTH;
+    }
+
+    return padding;
+  }
+
+  private getRightPadding(): number {
+    let padding = 0;
+
+    if (this.measureEntries.length === 1 && this.measureEntries[0] instanceof StaveSignature) {
+      padding += STAVE_SIGNATURE_ONLY_MEASURE_FRAGMENT_PADDING;
+    }
+
+    return padding;
   }
 }
